# sql/pgwire: embed commandResult in conn to avoid ClientComm allocations